// themes/GBA/BPGE/EmuDex/src/lib/types.ts
export interface Nature {
  name: string;
  /** Boosted stat index (0=Atk,1=Def,2=Spd,3=SpA,4=SpD). -1 = neutral */
  p: number;
  /** Nerfed stat index. -1 = neutral */
  m: number;
}

export interface StatBlock {
  hp: number;
  atk: number;
  def: number;
  spd: number;
  spatk: number;
  spdef: number;
}

export interface HiddenPower {
  type: string;
  power: number;
}

export interface DecryptedData {
  speciesId: number;
  heldItem: number;
  experience: number;
  friendship: number;
  moves: number[];
  pp: number[];
  evs: StatBlock;
  ivs: StatBlock;
  isEgg: boolean;
  abilityNum: number;
}

export interface Pokemon {
  pv: number;
  otid: number;
  speciesName: string;
  speciesId: number;
  level: number;
  hp: number;
  maxhp: number;
  atk: number;
  def: number;
  speed: number;
  spatk: number;
  spdef: number;
  status: number;
  nature: Nature;
  ivs: StatBlock;
  evs: StatBlock;
  moves: number[];
  pp: number[];
  friendship: number;
  hiddenPower: HiddenPower;
  isEgg: boolean;
  heldItem: number;
  ability: number;
}

export interface BattleMon {
  species: number;
  speciesName: string;
  atk: number;
  def: number;
  speed: number;
  spatk: number;
  spdef: number;
  moves: number[];
  ivs: StatBlock;
  pp: number[];
  hp: number;
  level: number;
  maxhp: number;
  type1: number;
  type2: number;
  status: number;
  pv: number;
  ability: number;
  item: number;
  statStages: number[];
  personality: number;
  friendship: number;
  status2: number;
}

export type WeatherType = 'none' | 'rain' | 'sun' | 'sand' | 'hail';

export interface SideState {
  reflect: number;
  lightScreen: number;
  safeguard: number;
  mist: number;
  spikes: number;
}

export interface FieldState {
  weather: WeatherType;
  weatherRaw: number;
  playerSide: SideState;
  enemySide: SideState;
}

export interface BattleState {
  active: boolean;
  isTrainer: boolean;
  isSafari: boolean;
  isDoubles: boolean;
  player: BattleMon | null;
  enemy: BattleMon | null;
  player2: BattleMon | null;
  enemy2: BattleMon | null;
  field: FieldState;
}

export interface SafariState {
  ballsLeft: number;
  stepsLeft: number;
  catchFactor: number;
  escapeFactor: number;
  rockCounter: number;
  baitCounter: number;
  inZone: boolean;
}
// ...
export interface MapState {
  mapGroup: number;
  mapNum: number;
  mapKey: string;
  mapName: string;
  playerGender: number; // 0 = male, 1 = female
}

export interface DexState {
  owned: Set<number>;
  seen: Set<number>;
}

export interface BagBall {
  itemId: number;
  quantity: number;
}